% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid}
\alias{covid}
\title{USA COVID-19 Data.}
\format{
A tibble with 127,232 rows and 7 variables:
\itemize{
\item \strong{id}: Index ID
\item \strong{date}:  Date when the cases are recorded
\item \strong{month}: Month when the cases are recorded
\item \strong{key_alpha_2}: State abbreviation
\item \strong{administrative_area_level_2}: USA States
\item \strong{variables}: Variable of interests (number of tests, confirmed cases, deaths, and recovered)
\item \strong{value}: The counts of the variable of interests.
}
}
\usage{
covid
}
\description{
This dataset contains USA COVID-19 data at State level from 2020-01-12 to 2020-10-21.
}
\examples{
library(tidyverse)
library(ggplot2)
data(covid)

# Creating a plot of cumulative number of COVID-19 cases in California overtime
covid \%>\% filter(variable == "confirmed", key_alpha_2 == "CA") \%>\% 
  ggplot(aes(x = date, y = value)) + 
  geom_line()

}
\references{
Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal of Open Source Software 5(51):2376, doi: 10.21105/joss.02376.
}
\keyword{datasets}
