% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{covid_filter}
\alias{covid_filter}
\title{Filtering the \code{covid} data frame based on user input.}
\usage{
covid_filter(data, stats)
}
\arguments{
\item{data}{\code{covid} data frame.}

\item{stats}{Preferred statistics to calculate the variable of interest in the COVID-19 data frame.
Use 'cumulative' to return cumulative COVID-19 stats, and 'daily'to return daily COVID-19 stats.}
}
\value{
\code{covid_filter} reruns the same type of the input, i.e. a data frame.
}
\description{
Filtering the \code{covid} data frame based on user input.
}
\examples{
# This will return filtered cumulative stats from covid data frame based on users' input
covid_filter(covid, stats = 'cumulative')

}
