% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{daterange}
\alias{daterange}
\title{Function to create date range input for Shiny UI with specified parameters}
\usage{
daterange(
  id,
  label = "Select date range",
  start = "2020-03-01",
  end = "2020-04-30",
  min = "2020-03-01",
  max = "2020-10-07"
)
}
\arguments{
\item{id}{Input ID that will be used to access the value chosen by users.}

\item{label}{The label for the date range input control.}

\item{start}{The initial start date. Either a Date object, or a string in yyyy-mm-dd format.}

\item{end}{The initial end date. Either a Date object, or a string in yyyy-mm-dd format.}

\item{min}{The minimum allowed date. Either a Date object, or a string in yyyy-mm-dd format.}

\item{max}{The maximum allowed date. Either a Date object, or a string in yyyy-mm-dd format.}
}
\value{
\code{daterange} returns A Date vector of length 2.
}
\description{
Function to create date range input for Shiny UI with specified parameters
}
\references{
Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson (2020). shiny: Web Application Framework for R. R
package version 1.5.0. https://CRAN.R-project.org/package=shiny
}
