% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{state_list}
\alias{state_list}
\title{USA State List.}
\format{
A dataframe with 49 rows and 3 variables:
\itemize{
\item \strong{X}: The value of the x axis for a state's centroid
\item \strong{Y}:The value of the y axis for a state's centroid
\item \strong{abb}: State abbreviation
}
}
\usage{
state_list

#'@references 
Pebesma, E., 2018. Simple Features for R: Standardized Support for Spatial Vector Data. The R Journal 10 (1), 439-446,
https://doi.org/10.32614/RJ-2018-009
}
\description{
This dataset contains a list of the centroids coordinates of each states in the US.
}
\keyword{datasets}
