% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{usa_map}
\alias{usa_map}
\title{USA Map Data.}
\format{
A dataframe with 15,537 rows and 10 variables:
\itemize{
\item \strong{long}: Longitude
\item \strong{lat}:  Latitude
\item \strong{group}: Grouping geographical data to states
\item \strong{order}: Index
\item \strong{region}: USA States
\item \strong{subregion}: NA
\item \strong{X}: The value of the x axis for a state's centroid
\item \strong{Y}:The value of the y axis for a state's centroid
\item \strong{geom}: Simple feature geometry list-column (sfc)
\item \strong{abb}: State abbreviation
}
}
\usage{
usa_map
}
\description{
This dataset contains necessary data to draw geographical map of the US.
}
\examples{
library(ggplot2)
data(usa_map)

# Creating a map of the US
ggplot(usa_map, aes(x = long, y = lat, group = group)) +
geom_polygon(color = "light grey",
            size = 0.25) +
 geom_text(aes(X, Y, 
               label = abb), 
           size = 3,
           color = "white",
           family = "sans") +
 theme_void()+
 theme(legend.position = "none")

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\keyword{datasets}
