% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{covid_filter}
\alias{covid_filter}
\title{Filtering the \code{covid} data frame based on user input.}
\usage{
covid_filter(data, stats, input)
}
\arguments{
\item{data}{a data frame, in this instance the \code{covid} data frame.}

\item{stats}{Preferred statistics to calculate the variable of interest in the COVID-19 data frame.
Use 'cumulative' to return cumulative COVID-19 stats, and 'daily' to return daily COVID-19 stats.}

\item{input}{Variable of interest chosen by users through Shiny app interactivity.}
}
\value{
\code{covid_filter} returns the same type of the input, i.e. a data frame.
}
\description{
Filtering the \code{covid} data frame based on user input.
}
